refactor(pos) :  패키지 - 테스트 패키지 변경
