+++ src/test/java/kitchenpos/menu/application/MenuGroupServiceTest.java
@@ -1,8 +1,7 @@
-package kitchenpos.application;
+package kitchenpos.menu.application;
 
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuGroupRepository;
-import kitchenpos.menu.application.MenuGroupService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
