+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -1,6 +1,7 @@
-package kitchenpos.application;
+package kitchenpos.menu.application;
 
-import kitchenpos.menu.application.MenuService;
+import kitchenpos.menu.infra.FakePurgomalumClient;
+import kitchenpos.product.application.InMemoryProductRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuProduct;
