+++ src/test/java/kitchenpos/order/common/application/InMemoryOrderRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.application;
+package kitchenpos.order.common.application;
 
 import kitchenpos.order.common.domain.Order;
 import kitchenpos.order.common.domain.OrderRepository;
